Serialized and Parallelized Particle Simulator
As can be accessed in the "Serial" and "Cuda" files, a particle simulator is implemented similarly via serial and parallel algorithms respectively. The serial algorithms include a sweep-and-prune, spatial hash, quadtree, and brute-force approach with appropriate tags to switch between them. Meanwhile, the parallel algorithms have a brute-force and sweep-and-prune algorithm fully implemented. My focus was building the quadtree, and I have found too many data dependencies to effectively parallelize it without severely modifying the whole program. However, I have learned a lot and enjoyed working with this team of three on a unique challenge for me.